Refactor + Rotation Input + Movement Input
Code was completely refactored so that the most used matrix and vector operations are in a separate area of the code. User input is also received to move the camera around and rotate it in a 'Wolfenstein' kind of behaviour. There is still triangle clipping to do and to correct the cameras Y axis direction.